refactor: 进一步拆分VirtualFileSystemPlugin模块，增强可维护性
