+++ src/plugins/VirtualFileSystemPlugin/challengeProcessor/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './collector';
+export * from './processor';
