+++ src/plugins/VirtualFileSystemPlugin/imageProcessor/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export * from './utils';
+export * from './markdownProcessor';
