+++ src/plugins/VirtualFileSystemPlugin/imageProcessor/types.ts
@@ -1,0 +1,6 @@
+export interface ImageProcessResult {
+    success: boolean;
+    data: string;
+    mimeType: string;
+    fullPath: string;
+}
